moved e listener from app.js to mycard
