+++ src/MyCard/contactInfo.json
@@ -1,0 +1,8 @@
+{
+  "firstName" : "Chiara",
+  "lastName" : "Marcial Martínez",
+  "title" : "full stack software engineer",
+  "email" : "contact@cofuente.io",
+  "githubUrl" : "https://github.com/cofuente",
+  "linkedInUrl" : "https://www.linkedin.com/in/cofuente/"
+}
